Upgrade to new task status view.
