+++ Launcher/Views/Content/CLTaskItemView.swift
@@ -8,28 +8,6 @@
 import Darwin
 import SwiftUI
 
-struct CLTaskStatusIndicatorView: View {
-    var status: CLTaskStatus = CLTaskStatus.stopped
-
-    var body: some View {
-        VStack {
-            switch status {
-                case .running:
-                    Circle()
-                .frame(width: 11, height: 11, alignment: .center)
-                .foregroundColor(Color.green)
-                case .failed:
-                    Circle()
-                .frame(width: 11, height: 11, alignment: .center)
-                .foregroundColor(Color.red)
-                case .stopped:
-                    Circle()
-                .frame(width: 11, height: 11, alignment: .center)
-                .foregroundColor(Color.gray)
-            }
-        }
-    }
-}
 
 struct CLTaskItemView: View {
     @EnvironmentObject private var store: CLStore
